requirements.txt
regex
google-genai
transformers
torch
